Extracted collectedTests functionality
