+++ scripts/test-reviewer/utils.mjs
@@ -1,0 +1,14 @@
+const unescapeHTML = string =>
+  string.replace(
+    /&amp;|&lt;|&gt;|&#39;|&quot;/g,
+    tag =>
+      ({
+        '&amp;': '&',
+        '&lt;': '<',
+        '&gt;': '>',
+        '&#39;': "'",
+        '&quot;': '"',
+      }[tag] || tag)
+  );
+
+export { unescapeHTML };
